refactor: 이전, 이후 버튼 분리
